% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheDir.R
\name{cacheDir}
\alias{cacheDir}
\title{Create a local cache directory for a package}
\usage{
cacheDir(pkg)
}
\arguments{
\item{pkg}{\code{character(1)}.
Package name.}
}
\value{
\code{character(1)}.
Directory path.
}
\description{
Create a local cache directory for a package
}
\details{
Requires the tools package to be installed.
}
\note{
Updated 2023-09-28.
}
\examples{
dir <- cacheDir(pkg = "AcidBase")
print(dir)
}
\seealso{
\itemize{
\item \code{\link[tools:userdir]{tools::R_user_dir()}}.
\item BiocFileCache package.
}
}
