% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classContains.R
\name{classContains}
\alias{classContains}
\title{Get the class contained in a specificed class definition}
\usage{
classContains(class, package)
}
\arguments{
\item{class}{\code{character(1)}.
Class name.}

\item{package}{\code{character(1)}.
Package name.}
}
\value{
\code{character(1)}.
Name of extended class.
}
\description{
Determine which class is contained (extended) in the requested class
definition.
}
\note{
Updated 2021-01-28.
}
\seealso{
\itemize{
\item \code{\link[methods:getClass]{methods::getClass()}}.
\item help("classRepresentation-class", "methods")
}
}
