% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerceToList-methods.R
\name{coerceToList}
\alias{coerceToList}
\alias{coerceToList,Annotated-method}
\title{Coerce to list}
\usage{
coerceToList(object, ...)

\S4method{coerceToList}{Annotated}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{list}.
}
\description{
Coerce to list
}
\note{
Don't attempt to defined methods for \code{as.list()}.

Updated 2021-10-14.
}
\examples{
data(DFrame, package = "AcidTest")

## Annotated ====
object <- DFrame
is(object, "Annotated")
object <- coerceToList(object)
class(object)
names(object)
}
\seealso{
Refer to the S4Vectors package for details.
\itemize{
\item \code{help(topic = "Annotated-class", package = "S4Vectors")}
\item \code{help(topic = "List-class", package = "S4Vectors")}
\item \code{showClass("Annnotated")}.
\item \code{showClass("List")}.
\item \code{methods::coerce}.
}
}
