% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathString.R
\name{collapseToPathString}
\alias{collapseToPathString}
\title{Collapse the input to a \code{PATH} string.}
\usage{
collapseToPathString(x)
}
\arguments{
\item{x}{\code{character}.
Elements to combine into a path string.}
}
\value{
\code{character(1)}.
}
\description{
Collapse the input to a \code{PATH} string.
}
\note{
Updated 2021-08-23.
}
\examples{
collapseToPathString(c("/usr/local/bin", "/usr/bin"))
}
