% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots}
\alias{dots}
\title{Extract dots from function}
\usage{
dots(..., character = FALSE)
}
\arguments{
\item{...}{Objects as dots.}

\item{character}{\code{logical(1)}.
Return dots (\code{...}) as \code{character}.}
}
\value{
\itemize{
\item \code{character = FALSE}: \code{list}.
Objects as \code{name} class.
Can return the object from the \code{name} with \code{eval}.
\item \code{character = TRUE}: \code{character}.
Names of the dots.
}
}
\description{
Extract dots from function
}
\note{
This function uses non-standard evaluation.

Updated 2020-01-28.
}
\examples{
dots(a, b, c, character = FALSE)
dots(a, b, c, character = TRUE)
}
\seealso{
\code{help("dotsMethods", "methods")}.
\itemize{
\item \href{https://rlang.r-lib.org/}{rlang}.
\item \href{https://www.tidyverse.org/}{tidyverse}.
}
}
