% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupes.R
\name{dupes}
\alias{dupes}
\title{Extract duplicated values}
\usage{
dupes(x)
}
\arguments{
\item{x}{\code{vector}.}
}
\value{
Modified object.
Only duplicated values are returned, sorted and made unique.
Returns empty vector of length 0 when no duplicates are detected.
}
\description{
Extract duplicated values
}
\note{
Updated 2023-09-26.
}
\examples{
## character ====
x <- c("ccc", "aaa", "bbb", "ccc", "bbb")
x <- dupes(x)
print(x)
}
\seealso{
\itemize{
\item \code{goalie::isDuplicate()}.
\item \code{duplicated()}.
}
}
