% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/euclidean-methods.R
\name{euclidean}
\alias{euclidean}
\alias{euclidean,numeric-method}
\title{Calculate Euclidean distance}
\usage{
euclidean(x, ...)

\S4method{euclidean}{numeric}(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
}
\description{
Calculate Euclidean distance
}
\note{
Updated 2023-09-11.
}
\examples{
## numeric ====
x <- seq(from = 1L, to = 5L, by = 1L)
y <- seq(from = 6L, to = 10L, by = 1L)
euclidean(x = x, y = y)
}
\seealso{
\itemize{
\item https://www.statology.org/euclidean-distance-in-r/
}
}
