% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDepth.R
\name{fileDepth}
\alias{fileDepth}
\title{File depth}
\usage{
fileDepth(..., fsep = .Platform[["file.sep"]])
}
\description{
This function is parameterized, and works on multiple files and/or
directories.
}
\details{
File path(s) do not have to exist on disk, similar to the approach employed
by \code{\link[base:file.path]{file.path()}}.

Note that "/" will return 0 here (zero indexed).
}
\note{
Updated 2019-09-05.
a <- tempdir()
b <- tempfile()
fileDepth(a, b)
}
