% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileExt.R
\name{fileExt}
\alias{fileExt}
\title{File extension}
\usage{
fileExt(path, pattern = goalie::extPattern)
}
\arguments{
\item{path}{\code{character}.
File path(s).
This function is vectorized and supports multiple files.}

\item{pattern}{\code{character(1)}.
File extension pattern.}
}
\value{
\code{character}.
Character vector of same length as \code{path} input, with file extension removed.
Returns \code{NA} if no extension is detected.
}
\description{
File extension
}
\note{
This function intentionally doesn't check whether a file exists.

Updated 2022-05-02.
}
\examples{
fileExt(c("dir/foo.txt", "dir/bar.tar.gz", "dir/"))
}
\seealso{
\code{tools::file_ext()}. Note that this returns \code{character()} instead of \code{NA}
on extension match failure.
}
