% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceDetach.R
\name{forceDetach}
\alias{forceDetach}
\title{Force detach packages}
\usage{
forceDetach(keep = .packages())
}
\arguments{
\item{keep}{\code{character}.
Package names to keep attached in environment.}
}
\value{
Invisible \code{logical(1)}.
}
\description{
Force detach unwanted packages loaded into the current environment.
}
\details{
Bioconductor ensembldb will attach unwanted packages into the NAMESPACE,
which can conflict with tidyverse packages (e.g. dplyr).
}
\note{
Updated 2022-05-20.
}
\examples{
## > forceDetach()
}
