% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{formalsList}
\alias{formalsList}
\title{List of global formals}
\format{
An object of class \code{list} of length 14.
}
\usage{
formalsList
}
\description{
These can be overwritten using the \code{options} function.
}
\note{
Updated 2020-01-19
}
\keyword{datasets}
