% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geometricMean-methods.R
\name{geometricMean}
\alias{geometricMean}
\alias{geometricMean,AsIs-method}
\alias{geometricMean,matrix-method}
\alias{geometricMean,numeric-method}
\title{Geometric mean}
\usage{
geometricMean(x, ...)

\S4method{geometricMean}{AsIs}(x)

\S4method{geometricMean}{matrix}(x, MARGIN = 2L)

\S4method{geometricMean}{numeric}(x, removeNA = TRUE, zeroPropagate = FALSE)
}
\arguments{
\item{x}{Object.}

\item{MARGIN}{\code{integer(1-2)}.
Dimension where the function will be applied.
For a two-dimensional matrix:
\code{1} indicates rows;
\code{2} indicates columns;
\code{c(1, 2)} indicates rows and columns.}

\item{removeNA}{\code{logical(1)}.
Remove \code{NA} values from calculations.}

\item{zeroPropagate}{\code{logical(1)}.
Allow propagation of zeroes.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
}
\description{
The geometric mean is the nth root of n products or e to the mean log of \code{x}.
}
\details{
Useful for describing non-normal (i.e. geometric) distributions.
}
\note{
This function should be fully zero- and \code{NA}-tolerant. This calculation
is not particularly useful if there are elements that are <= 0 and will
return \code{NaN}.

Updated 2021-10-14.
}
\examples{
## numeric ====
vec1 <- seq(from = 1L, to = 5L, by = 1L)
print(vec1)
geometricMean(vec1)

vec2 <- vec1^2L
print(vec2)
geometricMean(vec2)

## matrix ====
matrix <- matrix(
    data = c(vec1, vec2),
    ncol = 2L,
    byrow = FALSE
)
print(matrix)
geometricMean(matrix)
}
\seealso{
\itemize{
\item \href{https://stackoverflow.com/a/25555105}{Paul McMurdie's code}.
\item \code{psych::geometric.mean()}.
}
}
