% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/headtail-methods.R
\name{headtail}
\alias{headtail}
\alias{headtail,list-method}
\alias{headtail,vector-method}
\alias{headtail,matrix-method}
\alias{headtail,data.frame-method}
\title{Return the first and last parts of an object}
\usage{
headtail(x, ...)

\S4method{headtail}{list}(x, n = 2L)

\S4method{headtail}{vector}(x, n = 2L)

\S4method{headtail}{matrix}(x, n = 2L)

\S4method{headtail}{data.frame}(x, n = 2L)
}
\arguments{
\item{x}{Object.}

\item{n}{\code{integer(1)}.
Positive integer denoting the number of first and last items to include.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Applies to both rows and columns, enabling quick inspection during
interactive use.
}
\details{
Inspired by the \code{\link[base:print]{print()}} method for \code{DataFrame} class
objects.
}
\section{Functions}{
\itemize{
\item \code{headtail,list-method}: Paste collapse to a \code{character(1)}.

\item \code{headtail,matrix-method}: Show first and last rows.

\item \code{headtail,data.frame-method}: Same method as \code{matrix}.
}}

\note{
Updated 2020-10-07.
}
\examples{
data(mtcars, package = "datasets")

## data.frame ====
headtail(mtcars)
}
\seealso{
\itemize{
\item \code{utils::head()}, \code{utils::tail()}.
\item \code{cat()}.
}
}
