% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDir.R
\name{initDir}
\alias{initDir}
\title{Initialize directory}
\usage{
initDir(dir)
}
\arguments{
\item{dir}{\code{character}.
Directory path(s).}
}
\value{
\code{character}.
Directory path(s).
}
\description{
Checks and creates a directory recursively automatically.
}
\details{
Useful for setting up functions that require writes to a user-specified
directory.
}
\note{
Updated 2023-05-16.
}
\examples{
dir <- file.path(tempdir2(), c("aaa", "bbb"))
dir.exists(dir)
dir <- initDir(dir)
dir.exists(dir)
print(dir)
unlink2(dir)
}
