% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/intersectionMatrix-methods.R
\name{intersectionMatrix}
\alias{intersectionMatrix}
\alias{intersectionMatrix,DataFrame-method}
\alias{intersectionMatrix,data.frame-method}
\alias{intersectionMatrix,list-method}
\alias{intersectionMatrix,matrix-method}
\title{Intersection matrix}
\usage{
intersectionMatrix(object, ...)

\S4method{intersectionMatrix}{DataFrame}(object)

\S4method{intersectionMatrix}{data.frame}(object)

\S4method{intersectionMatrix}{list}(object)

\S4method{intersectionMatrix}{matrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Intersection matrix
}
\note{
Updated 2021-08-13.
}
\examples{
object <- list(
    "aaa" = c("a", "b", "c", "d", "e", "f"),
    "bbb" = c("b", "c", "d", "e", "f", "g"),
    "ccc" = c("c", "d", "e", "f", "g", "h")
)
mat <- intersectionMatrix(object)
rowSums(mat)
}
