% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/keepOnlyAtomicCols-methods.R
\name{keepOnlyAtomicCols}
\alias{keepOnlyAtomicCols}
\title{Calculate Euclidean distance}
\usage{
keepOnlyAtomicCols(object, ...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
Numeric vector.
}
\description{
Calculate Euclidean distance
}
\note{
Updated 2023-09-08.
}
\examples{
print("FIXME")
}
