% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorMinorVersion.R
\name{majorMinorVersion}
\alias{majorMinorVersion}
\alias{minorVersion}
\title{Major and minor version}
\usage{
majorMinorVersion(x)

minorVersion(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{numeric_version}.
}
\description{
Major and minor version
}
\note{
Updated 2023-01-30.
}
\examples{
majorMinorVersion(c("1.0.0", "2.1"))
}
