% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/matchNested-methods.R
\name{matchNested}
\alias{matchNested}
\alias{matchNested,character,DataFrame-method}
\alias{matchNested,character,data.frame-method}
\alias{matchNested,character,list-method}
\title{Match inside nested elements}
\usage{
matchNested(x, table, ...)

\S4method{matchNested}{character,DataFrame}(x, table)

\S4method{matchNested}{character,data.frame}(x, table)

\S4method{matchNested}{character,list}(x, table)
}
\arguments{
\item{x}{The values to be matched.}

\item{table}{The values to be matched against.
Applies across rows for \code{DataFrame} method.}

\item{...}{Additional arguments.}
}
\value{
\code{integer}.
A positional vector corresponding to values defined in \code{table} the same
size as \code{x}.
}
\description{
Match inside nested elements
}
\note{
Updated 2023-03-02.
}
\examples{
## list ====
table <- list(
    c("a", "aa", "aaa"),
    list(c("b", "bb", "bbb")),
    list(list(c("c", "cc", "ccc")))
)
matchNested(x = "ccc", table = table)

## data.frame ====
table <- data.frame(
    "V1" = c("a", "aa", "aaa"),
    "V2" = I(list(
        c("b", "bb", "bbb"),
        c("c", "cc", "ccc"),
        c("d", "dd", "ddd")
    )
)
matchNested(x = "ddd", table = table)

## DataFrame ====
## FIXME
}
