% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{metadataBlacklist}
\alias{metadataBlacklist}
\title{Sample metadata blacklist}
\format{
An object of class \code{character} of length 16.
}
\usage{
metadataBlacklist
}
\description{
Sample metadata blacklist
}
\note{
\code{sampleId} (\code{sampleID} for older datasets) is set automatically for
multiplexed/cell-level data.

Updated 2021-01-14.
}
\examples{
metadataBlacklist
}
\keyword{datasets}
