% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{metadataDenylist}
\alias{metadataDenylist}
\title{Sample metadata denylist}
\format{
An object of class \code{character} of length 16.
}
\usage{
metadataDenylist
}
\value{
\code{character}.
}
\description{
Sample metadata denylist
}
\details{
\code{sampleId} (\code{sampleID} for older datasets) is set automatically for
multiplexed / cell-level data.
}
\note{
Updated 2023-09-21.
}
\examples{
metadataDenylist
}
\keyword{datasets}
