% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{metricsCols}
\alias{metricsCols}
\title{Quality control metric columns}
\format{
An object of class \code{character} of length 6.
}
\usage{
metricsCols
}
\value{
\code{character}.
}
\description{
Column names returned by \code{calculateMetrics()}.
}
\details{
Previously: \code{"nGene"}, \code{"log10GenesPerUMI"}.
}
\note{
Updated 2023-09-21.
}
\examples{
metricsCols
}
\keyword{datasets}
