% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minorVersion.R
\name{minorVersion}
\alias{minorVersion}
\title{Minor version}
\usage{
minorVersion(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{numeric_version}.
}
\description{
Minor version
}
\note{
Updated 2020-04-09.
}
\examples{
minorVersion(c("1.0.0", "2.1"))
}
