% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentDirectory.R
\name{parentDirectory}
\alias{parentDirectory}
\alias{parentDir}
\title{Parent directory recursion}
\usage{
parentDirectory(path, n = 1L)

parentDir(path, n = 1L)
}
\arguments{
\item{path}{\code{character}.
Directory path(s).}

\item{n}{\code{logical(1)}.
Number of levels to recursive.
Defaults to the immediate parent, similar to \code{dirname}.}
}
\value{
\code{character}.
}
\description{
Parent directory recursion
}
\details{
Function always resolves directory path internally, using \code{realpath}.
}
\note{
Updated 2021-08-19.
}
\examples{
parentDir(path = getwd())
}
