% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseArgs.R
\name{parseArgs}
\alias{parseArgs}
\title{Parse command line argument flags}
\usage{
parseArgs(
  requiredFlags = NULL,
  optionalFlags = NULL,
  requiredArgs = NULL,
  optionalArgs = NULL,
  positionalArgs = FALSE
)
}
\arguments{
\item{requiredFlags, optionalFlags}{\code{character} or \code{NULL}.
Valid long flag names.
For example, \code{aaa} for \code{--aaa}.}

\item{requiredArgs, optionalArgs}{\code{character} or \code{NULL}.
Valid key-value pair argument names.
For example, \code{aaa} for \code{--aaa=AAA} or \verb{--aaa AAA}.
Note that \verb{--aaa AAA}-style arguments are not currently supported.}

\item{positional}{\code{logical(1)}.
Error if required positional arguments (>= 1) are missing.}
}
\value{
\code{list}.
Named list containing arguments, organized by type:
\itemize{
\item \code{requiredFlags}
\item \code{optionalFlags}
\item \code{requiredArgs}
\item \code{optionalArgs}
\item \code{positionalArgs}
}
}
\description{
Parse command line argument flags
}
\note{
Updated 2020-08-09.

Short flags, such as \code{-r}, are intentionally not supported.
}
\examples{
command <- system.file("scripts", "parse-args", package = "acidbase")
args <- c(
    ## Required flags:
    "--aaa", "--bbb",
    ## Optional flags:
    "--ccc", "--ddd",
    ## Required args:
    "--eee=EEE", "--fff=FFF",
    ## Optional args:
    "--ggg=GGG", "--hhh=HHH",
    ## Positional args:
    "III", "JJJ"
)
readLines(command)
out <- shell(command = command, args = args, stdout = TRUE)
print(out)
}
\seealso{
\itemize{
\item argparse Python package.
\item argparser R package.
\item optparse R package.
}
}
