% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteURL.R
\name{pasteURL}
\alias{pasteURL}
\title{Concatenate strings to form a URL}
\usage{
pasteURL(..., protocol = c("none", "https", "http", "ftp", "rsync", "s3"))
}
\arguments{
\item{...}{Character strings.
Intentionally does not support recycling here, unlike base \code{paste}.
Input of character vector as first argument alone is supported.
Errors if \code{NA} values are present.}

\item{protocol}{\code{character(1)}.
Desired protocol to use. Defaults to "https" but "http", "ftp", and "s3"
(AWS S3) are also supported. Use \code{"none"} if you want to prepare a URL that
already contains a protocol in the first element of the dots.}
}
\value{
\code{character(1)}.
URL path.
}
\description{
Concatenate strings to form a URL
}
\details{
Encoding is applied automatically via \code{utils::URLencode}.
}
\note{
Updated 2023-09-19.
}
\examples{
## HTTPS.
x <- pasteURL(
    "r.acidgenomics.com",
    "packages",
    "acidbase",
    protocol = "https"
)
print(x)

## FTP.
x <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-94",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.94.gtf.gz",
    protocol = "ftp"
)
print(x)

## Automatic encoding support.
x <- pasteURL(
    "rest.ensembl.org",
    "info",
    "assembly",
    "Homo sapiens",
    protocol = "https"
)
print(x)

## Character vector support.
vec <- c("r.acidgenomics.com", "packages", "acidbase")
x <- pasteURL(vec, protocol = "https")
print(x)
}
