% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalArgs.R
\name{positionalArgs}
\alias{positionalArgs}
\title{Positional command line arguments}
\usage{
positionalArgs(required = TRUE)
}
\arguments{
\item{required}{\code{logical(1)}.
Error if no positional arguments are defined.}
}
\value{
\code{character}.
}
\description{
These arguments do not contain \verb{--} prefixes.
}
\details{
Intended for using inside Rscript files.
}
\note{
Updated 2020-05-09.
}
\examples{
## $ Rscript test.Rscript "AAA" "BBB"
## > positionalArgs()
## c("AAA", "BBB")
}
