% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{Perform action quietly}
\usage{
quiet(expr)
}
\value{
Invisible \code{NULL}.
}
\description{
Suppress all warnings, messages, and console output.
}
\note{
Updated 2023-04-26.
}
\examples{
quiet({
    message("hello world")
})
}
