% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomString.R
\name{randomString}
\alias{randomString}
\title{Generate a random string}
\usage{
randomString(n = 10L)
}
\arguments{
\item{n}{\code{integer(1)}.
Number of characters to return. Length of string.}
}
\value{
\code{character(1)}.
}
\description{
Generate a random string
}
\note{
Updated 2023-08-01.
}
\examples{
x <- randomString()
print(x)
}
\seealso{
\itemize{
\item https://stackoverflow.com/questions/42734547/
}
}
