% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rankedMatrix-methods.R
\name{rankedMatrix}
\alias{rankedMatrix}
\alias{rankedMatrix,matrix-method}
\title{Calculate a ranked matrix}
\usage{
rankedMatrix(object, ...)

\S4method{rankedMatrix}{matrix}(
  object,
  MARGIN = 2L,
  method = c("increasing", "decreasing", "bidirectional"),
  tiesMethod = c("average", "first", "last", "random", "max", "min")
)
}
\arguments{
\item{object}{Object.}

\item{MARGIN}{\code{integer(1)}.
Direction to apply the function.
1 indicates rows, 2 indicates columns.}

\item{method}{\code{character(1)}.
Rank the values in \code{increasing}, \code{decreasing}, or \code{bidirectional} order.}

\item{tiesMethod}{\code{character(1)}.
How ties should be treated.
See \code{\link[=rank]{rank()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Calculate a ranked matrix
}
\details{
Ties are resolved automatically by calculating the average.
See the \code{ties.method} parameter in \code{\link[base:rank]{rank()}} for details.
}
\note{
Updated 2021-02-04.
}
\examples{
data(matrix_lfc, package = "AcidTest")
object <- matrix_lfc

## Increasing (negative to positive).
rankedMatrix(object, method = "increasing")

## Decreasing (positive to negative).
rankedMatrix(object, method = "decreasing")

## Bidirectional.
rankedMatrix(object, method = "bidirectional")
}
\seealso{
\itemize{
\item \code{\link[base:rank]{rank()}}.
\item Not to be confused with \code{Matrix::rankMatrix()}.
}
}
