% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{metadata}
\alias{metadata<-}
\alias{bapply}
\alias{compressExtPattern}
\alias{extPattern}
\alias{getNameInParent}
\alias{printString}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidGenerics}{\code{\link[AcidGenerics:reexports]{metadata}}, \code{\link[AcidGenerics:reexports]{metadata<-}}}

  \item{goalie}{\code{\link[goalie:engine-bapply]{bapply}}, \code{\link[goalie:extPattern]{compressExtPattern}}, \code{\link[goalie]{extPattern}}, \code{\link[goalie:engine-getNameInParent]{getNameInParent}}, \code{\link[goalie:engine-printString]{printString}}}
}}

