% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeVersion.R
\name{sanitizeVersion}
\alias{sanitizeVersion}
\title{Sanitize version}
\usage{
sanitizeVersion(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Sanitize version
}
\note{
Updated 2020-04-09.
}
\examples{
x <- c("2.7.15rc1", "1.10.0-patch1", "1.0.2k-fips")
sanitizeVersion(x)
}
