% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentenceCase.R
\name{sentenceCase}
\alias{sentenceCase}
\title{Sentence case}
\usage{
sentenceCase(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Sentence case
}
\note{
Updated 2019-10-21.
}
\examples{
x <- c("the quick Brown fox", "using AIC for model selection")
sentenceCase(x)
}
