% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/showHeader-methods.R
\name{showHeader}
\alias{showHeader}
\alias{showHeader,ANY-method}
\title{Header for object show method}
\usage{
showHeader(object, ...)

\S4method{showHeader}{ANY}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Console output, via \code{\link[base:cat]{cat()}}.
}
\description{
Header for object show method
}
\note{
Updated 2021-02-04.
}
\examples{
data(matrix, package = "AcidTest")
object <- matrix
showHeader(matrix)
}
\seealso{
\code{\link[=show]{show()}}.
}
