% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleClass.R
\name{simpleClass}
\alias{simpleClass}
\title{Simple class of object}
\usage{
simpleClass(x)
}
\value{
\code{character(1)}.
}
\description{
Simple class of object
}
\details{
Doesn't return \code{structure} information, including the parent package that
defines the class, which is common for S4 classes.

Primarily intended for unit testing in other packages.
}
\note{
Updated 2021-08-04.
}
\examples{
df <- data.frame()
simpleClass(df)
}
