% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathString.R
\name{splitPathString}
\alias{splitPathString}
\title{Split path string environment variable}
\usage{
splitPathString(x = Sys.getenv("PATH"))
}
\arguments{
\item{x}{\code{character(1)}.
PATH string, containing elements separated with a colon (\code{":"}).
Defaults to system path string.}
}
\value{
\code{character}.
Discrete path elements in a vector.
Intentionally does not make elements unique.
}
\description{
Split path string environment variable
}
\note{
Updated 2021-08-18.
}
\examples{
splitPathString()
}
\seealso{
\itemize{
\item \code{Sys.getenv("PATH")}
}
}
