% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdout.R
\name{stdout}
\alias{stdout}
\title{Split stdout string into a character vector}
\usage{
stdout(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{character}.
}
\description{
Split stdout string into a character vector
}
\note{
Updated 2022-05-02.
}
\examples{
x <- shell(
    command = "printf",
    args = c("\%s\n", "aaa", "bbb"),
    print = FALSE
)
print(x)
y <- stdout(x)
print(y)
}
\seealso{
\code{\link[=shell]{shell()}}.
}
