% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strExtract.R
\name{strExtract}
\alias{strExtract}
\title{Extract first pattern occurance}
\usage{
strExtract(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{\code{character}.
Character vector. \code{NA} values are allowed.}

\item{pattern}{\code{character(1)}.
Regular expression pattern.
Evalutes with \code{\link[=regexpr]{regexpr()}} internally.}

\item{fixed}{\code{logical(1)}.
If \code{TRUE}, \code{pattern} is a string to be matched as is.
Otherwise, will match by regular expression.}
}
\value{
\code{character}.
Character vector of matches. Failures return \code{NA}.
}
\description{
Uses \code{\link[=regexpr]{regexpr()}} internally without \code{\link[=regmatches]{regmatches()}}.
}
\details{
Expands with \code{NA} values for match failures, like stringi and stringr.
}
\note{
Updated 2023-09-25.
}
\examples{
object <- strExtract(
    x = c("apples x10", "bag of flour", "bag of sugar", "milk x200", NA),
    pattern = "\\\\d+"
)
print(object)
}
\seealso{
\itemize{
\item \code{\link[=regexpr]{regexpr()}}, \code{\link[=regmatches]{regmatches()}}.
\item \code{\link[stringi:stri_extract]{stringi::stri_extract_first()}}.
\item \code{\link[stringr:str_extract]{stringr::str_extract()}}.
\item https://stringr.tidyverse.org/articles/from-base.html
}
}
