% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strExtractAll.R
\name{strExtractAll}
\alias{strExtractAll}
\title{Extract all pattern occurances}
\usage{
strExtractAll(x, pattern, fixed = FALSE)
}
\description{
Extract all pattern occurances
}
\note{
Updated 2023-09-25.
}
\examples{
object <- strExtractAll(
    x= c("apples x4", "bag of flour", "bag of sugar", "milk x2"),
    pattern = "[a-z]+"
)
print(object)
}
\seealso{
\itemize{
\item \link[stringi:stri_extract]{stringi::stri_extract_all}.
\item \link[stringr:str_extract]{stringr::str_extract_all}.
\item https://stringr.tidyverse.org/articles/from-base.html
}
}
