% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripAttributes.R
\name{stripAttributes}
\alias{stripAttributes}
\title{Strip attributes}
\usage{
stripAttributes(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Strip attributes
}
\note{
Updated 2019-10-21.
}
\examples{
x <- "AAA"
attr(x, which = "test") <- "BBB"
x
attributes(x)
x <- stripAttributes(x)
attributes(x)
x
}
\seealso{
`assertive.base::strip_attributes().
}
