% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strMatch.R
\name{strMatch}
\alias{strMatch}
\title{Extract components from first match}
\usage{
strMatch(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{\code{character}.
Character vector. \code{NA} values are allowed.}

\item{pattern}{\code{character(1)}.
Regular expression pattern.
Evalutes with \code{\link[=regexec]{regexec()}} internally.}

\item{fixed}{\code{logical(1)}.
If \code{TRUE}, \code{pattern} is a string to be matched as is.
Otherwise, will match by regular expression.}
}
\value{
\code{matrix}.
Character matrix of match groups.
}
\description{
Uses \code{\link[=regexec]{regexec()}} and \code{\link[=regmatches]{regmatches()}} internally.
}
\details{
Expands with \code{NA} values for match failures, like stringi and stringr.
}
\note{
Updated 2023-09-25.
}
\examples{
## Regex match.
object <- strMatch(
    x = c("a-b", "c-d", "e_f", NA),
    pattern = "^(.+)-(.+)$",
    fixed = FALSE
)
print(object)

## Fixed match.
object <- strMatch(
    x = c("a", "aa", "b", "bb"),
    pattern = "a",
    fixed = TRUE
)
print(object)
}
\seealso{
\itemize{
\item \code{\link[=regexec]{regexec()}}, \code{\link[=regmatches]{regmatches()}}, \code{\link[=regexpr]{regexpr()}}.
\item \code{\link[stringi:stri_match]{stringi::stri_match_first()}}.
\item \code{\link[stringr:str_match]{stringr::str_match()}}.
\item https://stringr.tidyverse.org/articles/from-base.html
\item https://bookdown.org/rdpeng/rprogdatascience/regular-expressions.html
\item https://stackoverflow.com/questions/19171715/
\item https://d-rug.github.io/blog/2015/regex.fick
}
}
