% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strPad.R
\name{strPad}
\alias{strPad}
\title{Pad a string to minimum width}
\usage{
strPad(x, width, side, pad)
}
\arguments{
\item{x}{\code{character}.}
}
\description{
Pad a string to minimum width
}
\note{
Updated 2023-09-21.
}
\seealso{
\itemize{
\item \code{\link[stringr:str_pad]{stringr::str_pad()}}.
\item \code{\link[=sprintf]{sprintf()}}.
\item https://stringr.tidyverse.org/articles/from-base.html
}
}
