% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit2.R
\name{strsplit2}
\alias{strsplit2}
\title{Split the elements of a character vector into a matrix}
\usage{
strsplit2(x, split)
}
\value{
\code{matrix}.
}
\description{
Split the elements of a character vector into a matrix
}
\note{
Updated 2023-09-13.
}
\examples{
x <- c(
    "TSPAN6_ENST00000373020",
    "TSPAN6_ENST00000494424",
    "TSPAN6_ENST00000496771",
    "TSPAN6_ENST00000612152",
    "TSPAN6_ENST00000614008"
)
x <- strsplit2(x = x, split = "_")
print(x)
}
\seealso{
\itemize{
\item stringi::stri_split_fixed
\item stringr::str_split_fixed
}
}
