% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdir2.R
\name{tempdir2}
\alias{tempdir2}
\title{Unique temporary directory}
\usage{
tempdir2()
}
\value{
\code{character(1L)}.
Temporary directory path.
}
\description{
Unique temporary directory
}
\note{
Updated 2022-06-02.
}
\examples{
x <- tempdir2()
print(x)
}
