% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{truncate}
\alias{truncate}
\title{Truncate input to a specified width}
\usage{
truncate(x, width = getOption("width"))
}
\arguments{
\item{x}{Object.}

\item{width}{\code{integer(1)}.
Width.}
}
\description{
Automatically generate a substring and add trailing "..." if necessary.
}
\details{
Note that return matches the desired width.
}
\note{
Updated 2019-10-21.
}
\examples{
x <- "the quick brown fox"
x <- truncate(x, width = 10L)
nchar(x)
}
\seealso{
\code{assertive.base:::truncate()}.
}
