% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateString.R
\name{truncateString}
\alias{truncateString}
\title{Truncate a string with ellipsis, if necessary}
\usage{
truncateString(x, n = 200L)
}
\arguments{
\item{x}{\code{character(1)}.
Character string.}

\item{n}{\code{integer(1)}.
Maximum number of characters.
Function will truncate with \code{"..."}, when applicable.}
}
\value{
\code{character(1)}.
}
\description{
Truncate a string with ellipsis, if necessary
}
\note{
Updated 2023-09-21.
}
\examples{
truncateString("abcdefgi", n = 6L)
}
\seealso{
\itemize{
\item https://stackoverflow.com/questions/46759358/
}
}
