% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeDescription.R
\name{typeDescription}
\alias{typeDescription}
\title{Type description}
\usage{
typeDescription(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Type description
}
\note{
Updated 2019-10-21.
}
\examples{
typeDescription(NULL)
typeDescription(TRUE)
typeDescription(character())
typeDescription(data.frame())
}
