% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathString.R
\name{uniquePathString}
\alias{uniquePathString}
\title{Unique path string}
\usage{
uniquePathString(x = Sys.getenv("PATH"))
}
\arguments{
\item{x}{\code{character(1)}.
Path string, containing elements separated with a colon (\code{":"}).
Defaults to system path string.}
}
\value{
\code{character(1)}.
}
\description{
Unique path string
}
\note{
Updated 2021-08-17.
}
\examples{
uniquePathString()
}
\seealso{
\itemize{
\item \code{Sys.getenv("PATH")}
}
}
