% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlink2.R
\name{unlink2}
\alias{unlink2}
\title{Delete files and directories}
\usage{
unlink2(x)
}
\arguments{
\item{x}{\code{character}.
Files or directories to be deleted.
Unlink base \code{unlink}, these must exist on disk or the function will
intentionally error.}
}
\value{
\code{integer(1)}.
\code{0} for success, \code{1} for failure, invisibly.
}
\description{
Delete files and directories
}
\details{
This variant hardens file path handling, for better Windows compatibility.
}
\note{
Updated 2022-06-01.
}
\examples{
tempdir <- tempdir2()
x <- file.path(tempdir, "file.txt")
y <- file.path(tempdir, "directory")
invisible(file.create(x))
invisible(dir.create(y))
out <- unlink2(c(x, y))
print(out)
}
