% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{updateMessage}
\alias{updateMessage}
\title{Update message}
\format{
An object of class \code{character} of length 1.
}
\usage{
updateMessage
}
\value{
\code{character(1)}.
}
\description{
Update message
}
\note{
Updated 2023-09-21.
}
\examples{
message(updateMessage)
}
\keyword{datasets}
