% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zscore-methods.R
\name{zscore}
\alias{zscore}
\alias{zscore,matrix-method}
\alias{zscore,numeric-method}
\title{Calculate Z-score}
\usage{
zscore(object, ...)

\S4method{zscore}{matrix}(object, MARGIN = 1L)

\S4method{zscore}{numeric}(object)
}
\arguments{
\item{object}{Object.}

\item{MARGIN}{\code{integer(1-2)}.
Dimension where the function will be applied.
For a two-dimensional matrix:
\code{1} indicates rows;
\code{2} indicates columns;
\code{c(1, 2)} indicates rows and columns.}

\item{...}{Additional arguments.}
}
\value{
Numeric vector or matrix.
}
\description{
Calculate Z-score
}
\note{
Updated 2022-11-17.
}
\examples{
data(matrix, package = "AcidTest")

## numeric ====
object <- matrix[1L, ]
zscore(object)

## matrix ====
object <- matrix
zscore(matrix, MARGIN = 1L)
}
\seealso{
\itemize{
\item \code{scale}.
}
}
