% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidBase-package}
\alias{AcidBase}
\alias{AcidBase-package}
\title{AcidBase}
\description{
Low-level base functions imported by Acid Genomics packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://acidbase.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/AcidBase/}
  \item Report bugs at \url{https://github.com/acidgenomics/AcidBase/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
