% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize}
\usage{
capitalize(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Modified object, with names capitalized.
}
\description{
Capitalize
}
\note{
Updated 2019-10-21.
}
\examples{
capitalize(c("hello", "world"))
}
