% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\alias{sentenceCase}
\title{Capitalize}
\usage{
capitalize(x)

sentenceCase(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Modified object, with names capitalized.
}
\description{
Apply simple capitlization rules.
}
\details{
For complex renaming, refer to syntactic package.
}
\section{Functions}{
\itemize{
\item \code{capitalize}: Capitalize the first letter of all words.

\item \code{sentenceCase}: Only capitalize the first letter in the sentence.
}}

\note{
Updated 2019-10-21.
}
\examples{
x <- c("the quick brown fox", "using AIC for model selection")
capitalize(x)
sentenceCase(x)
}
\seealso{
\itemize{
\item \code{toupper()}, \code{tolower()}.
\item stringr package.
\item \code{tools::toTitleCase()}. I generally avoid title case, since there's a poor
general consensus on which works to keep lower case.
}
}
