% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpus.R
\name{cpus}
\alias{cpus}
\title{Detect the number of CPU cores}
\usage{
cpus()
}
\value{
\code{integer(1)}.
Number of CPU cores.
}
\description{
Detect the number of CPU cores
}
\note{
Updated 2023-05-16.
}
\examples{
cpus()
}
\seealso{
\itemize{
\item parallel::detectCores
\item ps::ps_cpu_count
}
}
