% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDepth.R
\name{fileDepth}
\alias{fileDepth}
\title{File depth}
\usage{
fileDepth(path, fsep = .Platform[["file.sep"]])
}
\arguments{
\item{path}{\code{character}.
File path(s).
This function is vectorized and supports multiple files.}

\item{fsep}{\code{character(1)}.
Platform-specific file path separator (e.g. "/" on Unix).}
}
\value{
\code{integer}.
Unnamed vector denoting file path depth.
Note that "/" will return 0 here (zero indexed).
}
\description{
File depth
}
\note{
This function intentionally doesn't check whether a file exists.

Updated 2019-09-05.
}
\examples{
a <- tempdir()
b <- tempfile()
fileDepth(a, b)
}
