% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNested.R
\name{filterNested}
\alias{filterNested}
\title{Filter a nested data frame by a matching pattern across all columns.}
\usage{
filterNested(object, pattern, ignoreCase = FALSE)
}
\arguments{
\item{object}{Object.}

\item{pattern}{\code{character(1)}.
Pattern to match rowwise, across all columns.}

\item{ignoreCase}{\code{logical(1)}.
Ignore case in pattern match.}

\item{fixed}{\code{logical(1)}.
Perform a fixed instead of regular expression match.}
}
\value{
Modified object, containing only matching rows.
}
\description{
Filter a nested data frame by a matching pattern across all columns.
}
\note{
Updated 2023-12-12.
}
