% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{formalsList}
\alias{formalsList}
\title{Shared list of optional default formals}
\format{
An object of class \code{list} of length 25.
}
\usage{
formalsList
}
\description{
These can be overwritten using the \code{options} function.
}
\note{
Updated 2021-09-09.
}
\examples{
head(formalsList)
}
\seealso{
\itemize{
\item https://ggplot2.tidyverse.org/reference/scale_colour_continuous.html
\item https://ggplot2.tidyverse.org/reference/scale_colour_discrete.html
}
}
\keyword{datasets}
