% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDir.R
\name{initDir}
\alias{initDir}
\title{Initialize directory}
\usage{
initDir(dir)
}
\arguments{
\item{dir}{\code{character(1)}.
Directory path.}
}
\value{
\code{character(1)}.
Directory path.
}
\description{
Checks and creates a directory recursively automatically.
}
\details{
Useful for setting up functions that require writes to a user-specified
directory.
}
\note{
Updated 2022-06-02.
}
\examples{
dir <- file.path(tempdir(), "testdir")
dir.exists(dir)
dir <- initDir(dir)
dir.exists(dir)
print(dir)
unlink2(dir)
}
