% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/keepOnlyAtomicCols-methods.R
\name{keepOnlyAtomicCols}
\alias{keepOnlyAtomicCols}
\alias{keepOnlyAtomicCols,DFrame-method}
\alias{keepOnlyAtomicCols,data.frame-method}
\title{Keep only atomic columns}
\usage{
keepOnlyAtomicCols(object, ...)

\S4method{keepOnlyAtomicCols}{DFrame}(object)

\S4method{keepOnlyAtomicCols}{data.frame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, containing only atomic columns.
}
\description{
Keep only atomic columns
}
\note{
Updated 2023-09-08.
}
\examples{
## DFrame ====
object <- S4Vectors::DataFrame(
    "a" = I(list(
        c(1L, 2L),
        c(3L, 4L),
        c(5L, 6L)
    )),
    "b" = c(TRUE, TRUE, FALSE),
    "c" = c("a", "b", "c"),
    row.names = c("x", "y", "z")
)
print(object)
object <- keepOnlyAtomicCols(object)
print(object)
}
