% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/keepOnlyAtomicCols-methods.R
\name{keepOnlyAtomicCols}
\alias{keepOnlyAtomicCols}
\alias{keepOnlyAtomicCols,DFrame-method}
\alias{keepOnlyAtomicCols,data.frame-method}
\title{Keep only atomic columns}
\usage{
keepOnlyAtomicCols(object, ...)

\S4method{keepOnlyAtomicCols}{DFrame}(object)

\S4method{keepOnlyAtomicCols}{data.frame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, containing only atomic columns.
}
\description{
Keep only atomic columns
}
\note{
Updated 2023-09-08.
}
\examples{
print("FIXME")
}
