% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorVersion.R
\name{majorVersion}
\alias{majorVersion}
\title{Major version}
\usage{
majorVersion(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Major version
}
\note{
Updated 2020-04-09.
}
\examples{
majorVersion("1.0.0")
}
