% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{metadataDenylist}
\alias{metadataDenylist}
\alias{metadataBlacklist}
\title{Sample metadata denylist}
\format{
An object of class \code{character} of length 16.
}
\usage{
metadataDenylist
}
\description{
Sample metadata denylist
}
\note{
\code{sampleId} (\code{sampleID} for older datasets) is set automatically for
multiplexed/cell-level data.

Updated 2021-02-25.
}
\examples{
metadataDenylist
}
\keyword{datasets}
