% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentDirectory.R
\name{parentDirectory}
\alias{parentDirectory}
\alias{parentDir}
\title{Parent directory recursion}
\usage{
parentDirectory(path, n = 1L, fsep = .Platform$file.sep)

parentDir(path, n = 1L, fsep = .Platform$file.sep)
}
\arguments{
\item{path}{\code{character}.
Local file and/or directory paths.
Also supports URLs.}

\item{n}{\code{logical(1)}.
Number of levels to recursive.
Defaults to the immediate parent, similar to \code{dirname}.}

\item{fsep}{\code{character(1)}.
Path separator to use (e.g. \code{"/"}).}
}
\value{
\code{character}.
}
\description{
Parent directory recursion
}
\details{
Function always resolves directory path internally, using \code{realpath}.
}
\note{
Updated 2021-09-03.
}
\examples{
parentDir(path = getwd())
}
