% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseArgs.R
\name{parseArgs}
\alias{parseArgs}
\title{Parse command line argument flags}
\usage{
parseArgs(positional = FALSE, validArgs = NULL, validFlags = NULL)
}
\arguments{
\item{positional}{\code{logical(1)}.
Error if required positional arguments are missing.}

\item{validArgs}{\code{character} or \code{NULL}.
Valid key-value pair argument names.
For example, \code{aaa} for \code{--aaa=AAA}.}

\item{validFlags}{\code{character} or \code{NULL}.
Valid flag names.
For example, \code{aaa} for \code{--aaa}.}
}
\value{
\code{character}.
Arguments.
}
\description{
Parse command line argument flags
}
\note{
Updated 2020-04-07.
}
\examples{
parseArgs()

## $ Rscript test.Rscript --a="A" --b
## > parseArgs(positional = FALSE, validArgs = "a", validFlags = "b")
}
\seealso{
\itemize{
\item argparser R package
\item argparse Python package
}
}
