% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteURL.R
\name{pasteURL}
\alias{pasteURL}
\title{Concatenate strings to form a URL}
\usage{
pasteURL(..., protocol = c("none", "https", "http", "ftp", "rsync", "s3"))
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{protocol}{\code{character(1)}.
Desired protocol to use. Defaults to "https" but "http", "ftp", and "s3"
(AWS S3) are also supported. Use \code{"none"} if you want to prepare a URL that
already contains a protocol in the first element of the dots.}
}
\value{
\code{character}.
URL path, containing forward slashes (\code{/}).
}
\description{
Concatenate strings to form a URL
}
\details{
Encoding is applied automatically via \code{utils::URLencode}.
}
\note{
Updated 2023-07-27.
}
\examples{
## HTTPS.
x <- pasteURL(
    "steinbaugh.com",
    "basejump",
    "reference",
    protocol = "https"
)
print(x)

## FTP.
x <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-94",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.94.gtf.gz",
    protocol = "ftp"
)
print(x)

## Automatic encoding support.
x <- pasteURL(
    "rest.ensembl.org",
    "info",
    "assembly",
    "Homo sapiens",
    protocol = "https"
)
print(x)
}
