% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalArgs.R
\name{positionalArgs}
\alias{positionalArgs}
\title{Positional command line arguments}
\usage{
positionalArgs()
}
\description{
These arguments do not contain \verb{--} prefixes.
}
\details{
Intended for using inside Rscript files.
}
\note{
Updated 2020-04-07.
}
\examples{
positionalArgs()
}
