% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram}
\alias{ram}
\title{Return system random access memory (RAM) information}
\usage{
ram(type = c("total", "free"), units = "GB")
}
\value{
\code{numeric(1)}.
}
\description{
Return system random access memory (RAM) information
}
\note{
Updated 2023-05-10.
}
\examples{
ram(type = "total", units = "GB")
ram(type = "free", units = "GB")
}
\seealso{
\itemize{
\item memuse package.
\item https://stackoverflow.com/questions/29787452/
\item \code{help("object.size")}
}
}
