% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rankedMatrix-methods.R
\name{rankedMatrix}
\alias{rankedMatrix}
\alias{rankedMatrix,matrix-method}
\title{Calculate a ranked matrix}
\usage{
showHeader(rankedMatrix, ...)

\S4method{rankedMatrix}{matrix}(
  object,
  MARGIN = 2L,
  method = c("increasing", "decreasing", "bidirectional")
)
}
\arguments{
\item{object}{Object.}

\item{method}{\code{character(1)}.
Rank the values in \code{increasing}, \code{decreasing}, or \code{bidirectional} order.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Calculate a ranked matrix
}
\details{
Ties are resolved automatically by calculating the average.
See the \code{ties.method} parameter in \code{\link[base:rank]{rank()}} for details.
}
\note{
Updated 2021-02-04.
}
\examples{
data(matrix_lfc, package = "AcidTest")
lfc <- matrix_lfc

## Increasing (negative to positive).
rankedMatrix(lfc, method = "increasing")

## Decreasing (positive to negative).
rankedMatrix(lfc, method = "decreasing")

## Bidirectional.
rankedMatrix(lfc, method = "bidirectional")
}
\seealso{
\itemize{
\item \code{\link[base:rank]{rank()}}.
\item Not to be confused with \code{Matrix::rankMatrix()}.
}
}
