% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realpath.R
\name{realpath}
\alias{realpath}
\title{Express file paths in canonical form}
\usage{
realpath(path)
}
\arguments{
\item{path}{\code{character}.
File paths, which must exist on disk.}
}
\description{
Express file paths in canonical form
}
\note{
Updated 2022-05-31.

Use \code{normalizePath()} instead for non-existing file paths.
}
\examples{
realpath(".")
}
\seealso{
Standard path modifiers:
\itemize{
\item \code{normalizePath()}.
\item \code{file.path()}.
}

Checking for existence, access:
\itemize{
\item \code{file.access()}.
\item \code{file.exists()}.
\item \code{goalie::hasAccess()}.
\item \code{goalie::allHaveAccess()}.
}
}
