% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{URLencode}
\alias{arrow}
\alias{as.formula}
\alias{axisTicks}
\alias{bapply}
\alias{capture.output}
\alias{colorRampPalette}
\alias{compressExtPattern}
\alias{data}
\alias{download.file}
\alias{dist}
\alias{ecdf}
\alias{extPattern}
\alias{formula}
\alias{getNameInParent}
\alias{getS3method}
\alias{hclust}
\alias{head}
\alias{install.packages}
\alias{median}
\alias{metadata}
\alias{metadata<-}
\alias{model.matrix}
\alias{packageName}
\alias{packageVersion}
\alias{prcomp}
\alias{quantile}
\alias{relevel}
\alias{reorder}
\alias{rgb}
\alias{sessionInfo}
\alias{tail}
\alias{unit}
\alias{untar}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidGenerics}{\code{\link[AcidGenerics]{head}}, \code{\link[AcidGenerics:reexports]{median}}, \code{\link[AcidGenerics:reexports]{metadata}}, \code{\link[AcidGenerics:reexports]{metadata<-}}, \code{\link[AcidGenerics]{tail}}}

  \item{goalie}{\code{\link[goalie]{bapply}}, \code{\link[goalie:extPattern]{compressExtPattern}}, \code{\link[goalie]{extPattern}}, \code{\link[goalie]{getNameInParent}}}

  \item{grDevices}{\code{\link[grDevices]{axisTicks}}, \code{\link[grDevices:colorRamp]{colorRampPalette}}, \code{\link[grDevices]{rgb}}}

  \item{grid}{\code{\link[grid]{arrow}}, \code{\link[grid]{unit}}}

  \item{stats}{\code{\link[stats:formula]{as.formula}}, \code{\link[stats]{dist}}, \code{\link[stats]{ecdf}}, \code{\link[stats]{formula}}, \code{\link[stats]{hclust}}, \code{\link[stats]{model.matrix}}, \code{\link[stats]{prcomp}}, \code{\link[stats]{quantile}}, \code{\link[stats]{relevel}}, \code{\link[stats:reorder.factor]{reorder}}}

  \item{utils}{\code{\link[utils]{URLencode}}, \code{\link[utils]{capture.output}}, \code{\link[utils]{data}}, \code{\link[utils]{download.file}}, \code{\link[utils]{getS3method}}, \code{\link[utils]{install.packages}}, \code{\link[utils]{packageName}}, \code{\link[utils:packageDescription]{packageVersion}}, \code{\link[utils]{sessionInfo}}, \code{\link[utils]{untar}}}
}}

