% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sem-methods.R
\name{sem}
\alias{sem}
\alias{sem,numeric-method}
\alias{sem,AsIs-method}
\title{Standard error of the mean}
\usage{
sem(x, ...)

\S4method{sem}{numeric}(x)

\S4method{sem}{AsIs}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
}
\description{
Standard error of the mean
}
\details{
Alternatively, can use: \code{sd(x) / sqrt(length(x))}.
}
\note{
Updated 2020-11-24.
}
\examples{
x <- seq(from = 1L, to = 10L, by = 1L)
x <- sem(x)
x
}
\seealso{
\itemize{
\item https://stackoverflow.com/questions/2676554/
}
}
