% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shell}
\alias{shell}
\title{Invoke a command in the system command-line shell}
\usage{
shell(
  command,
  args = character(),
  print = interactive(),
  wd = getwd(),
  stdoutFile = NULL,
  stderrFile = NULL
)
}
\arguments{
\item{command}{\code{character(1)}.
Name of program to run.}

\item{args}{\code{character}.
Arguments passed to \code{command}.}

\item{print}{\code{logical(1)}.
Whether to print (echo) the commands to the console.}

\item{wd}{\code{character(1)}.
Working directory path inside shell session.}

\item{stdoutFile, stderrFile}{\code{character(1)} or \code{NULL}.
File path to log stdout and/or stderr.
Disabled when set \code{NULL}.}
}
\value{
Invisible \code{list}.
Contains named elements:
\code{"status"}, \code{"stdout"}, \code{"stderr"}, \code{"timeout"}.
}
\description{
Invoke a command in the system command-line shell
}
\note{
Updated 2021-08-18.
}
\examples{
x <- shell(
    command = "echo",
    args = c("hello", "world"),
    print = TRUE
)
print(x)
}
\seealso{
\itemize{
\item \code{processx::run()}.
\item \code{base::system2()}, our previously used legacy approach.
}
}
