% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shell}
\alias{shell}
\title{Invoke a command in the system command-line shell}
\usage{
shell(command, args = "", stdout = "", stderr = "", ...)
}
\arguments{
\item{command}{\code{character(1)}.}

\item{args}{\code{character}.}

\item{stdout, stderr}{\code{character(1)}, \code{logical(1)}, or \code{NULL}.
Where output to \code{stdout} or \code{stderr} should be sent.

Possible values:
\itemize{
\item \code{""}: Output to the R console.
\item \code{FALSE} or \code{NULL}: Discard the output.
\item \code{TRUE}: Capture the output in a character vector.
\item \code{character(1)}: Capture the output in a defined file name.
}}

\item{...}{Passthrough arguments to \code{\link[base:system2]{system2()}}.}
}
\value{
If \code{stdout = TRUE} or \code{stderr = TRUE}, a character vector giving the output
of the command, one line per character string. If the command could not be
run, an R error is generated. If command runs but gives a non-zero exit
status this will be reported with a warning and in the attribute "status" of
the result.

In other cases, the return value is an invisible error code
(\code{0} for success). If the command could not be run for any reason, the value
is \code{127} and a warning is issued. If the command times out, a warning is
issued and the exit status is \code{124}.
}
\description{
Invoke a command in the system command-line shell
}
\note{
Updated 2021-01-06.
}
\examples{
shell(command = "echo", args = c("hello", "world"))
}
\seealso{
\itemize{
\item \code{system2}.
}
}
