% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strMatch.R
\name{strMatch}
\alias{strMatch}
\title{Extract components from a match}
\usage{
strMatch(x, pattern)
}
\arguments{
\item{x}{\code{character}.
Character vector. \code{NA} values are allowed.}

\item{pattern}{\code{character(1)}.
Regular expression pattern.
Evalutes with \code{\link[=regexec]{regexec()}} internally.}
}
\value{
\code{matrix}.
Character matrix of match groups.
}
\description{
Uses \code{\link[=regexec]{regexec()}} and \code{\link[=regmatches]{regmatches()}} from base R internally.
}
\details{
Expands with \code{NA} values for match failures, like stringi and stringr.
}
\note{
Updated 2023-09-21.
}
\examples{
x <- c("a-b", "c-d", "e_f", NA)
pattern <- "^(.+)-(.+)$"
mat <- strMatch(x = x, pattern = pattern)
print(mat)
}
\seealso{
\itemize{
\item \code{\link[=regexec]{regexec()}}, \code{\link[=regmatches]{regmatches()}}, \code{\link[=regexpr]{regexpr()}}.
\item \code{\link[stringi:stri_match]{stringi::stri_match()}}.
\item \code{\link[stringr:str_match]{stringr::str_match()}}.
\item https://stringr.tidyverse.org/articles/from-base.html
\item https://bookdown.org/rdpeng/rprogdatascience/regular-expressions.html
\item https://stackoverflow.com/questions/19171715/
\item https://d-rug.github.io/blog/2015/regex.fick
}
}
