% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strMatchAll.R
\name{strMatchAll}
\alias{strMatchAll}
\title{Extract components from all matches}
\usage{
strMatchAll(x, pattern, fixed = FALSE)
}
\arguments{
\item{x}{\code{character}.
Character vector. \code{NA} values are allowed.}

\item{pattern}{\code{character(1)}.
Regular expression pattern.
Evalutes with \code{\link[=gregexec]{gregexec()}} internally.}

\item{fixed}{\code{logical(1)}.
If \code{TRUE}, \code{pattern} is a string to be matched as is.
Otherwise, will match by regular expression.}
}
\value{
\code{list}.
List of character matrix of match groups.
}
\description{
Uses \code{\link[=gregexec]{gregexec()}} and \code{\link[=regmatches]{regmatches()}} internally.
}
\details{
Expands with \code{NA} values for match failures, like stringi and stringr.
}
\note{
Updated 2023-09-25.
}
\examples{
## Regex match.
object <- strMatchAll(
    x = c("a-b", "c-d", "e_f", NA),
    pattern = "^(.+)-(.+)$",
    fixed = FALSE
)
print(object)

## Fixed match.
object <- strMatchAll(
    x = c("a", "aa", "b", "bb"),
    pattern = "a",
    fixed = TRUE
)
print(object)
}
\seealso{
\itemize{
\item \code{\link[=gregexec]{gregexec()}}, \code{\link[=regmatches]{regmatches()}}, \code{\link[=gregexpr]{gregexpr()}}.
\item \code{\link[stringi:stri_match]{stringi::stri_match_all()}}.
\item \code{\link[stringr:str_match]{stringr::str_match_all()}}.
\item https://stringr.tidyverse.org/articles/from-base.html
}
}
