% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strPad.R
\name{strPad}
\alias{strPad}
\title{Pad a string to minimum width}
\usage{
strPad(x, width, side = c("left", "right"), pad = " ")
}
\arguments{
\item{x}{\code{character}.
Character vector that does not contain \code{NA} or empty strings.}

\item{width}{\code{integer(1)}.
Padding width.}

\item{side}{\code{character(1)}.
Apply padding to the left or right side.}

\item{pad}{\code{character(1)}.
Single padding character.}
}
\value{
\code{character}.
}
\description{
Pad a string to minimum width
}
\note{
Updated 2023-09-21.
}
\examples{
x <- c("1", "10", "100")
width <- max(nchar(x))
strPad(x = x, width = width, side = "left", pad = "0")

}
\seealso{
\itemize{
\item \code{\link[stringr:str_pad]{stringr::str_pad()}}.
\item \code{\link[=sprintf]{sprintf()}}.
\item https://stringr.tidyverse.org/articles/from-base.html
}
}
