% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strReplaceNA.R
\name{strReplaceNA}
\alias{strReplaceNA}
\title{Replace \code{NA} with \code{"NA"}}
\usage{
strReplaceNA(x)
}
\arguments{
\item{x}{\code{character}.}
}
\value{
\code{character}
Character vector with \code{NA} values converted to \code{"NA"}.
}
\description{
Replace \code{NA} with \code{"NA"}
}
\note{
Updated 2023-09-21.
}
\examples{
strReplaceNA(c("a", "b", NA))
}
\seealso{
\itemize{
\item \link[stringi:stri_replace_na]{stringi::stri_replace_na}.
\item \link[stringr:str_replace_na]{stringr::str_replace_na}.
}
}
