% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strSplit.R
\name{strSplit}
\alias{strSplit}
\title{Split the elements of a character vector into a matrix}
\usage{
strSplit(x, split)
}
\arguments{
\item{x}{\code{character}.
Character vector that does not contain \code{NA} or empty strings.}

\item{split}{\code{character(1)}.
String that defines the split boundary.
Regular expressions are intentionally not supported to keep this simple.}
}
\value{
\code{matrix}.
Character matrix split into columns.
}
\description{
Split the elements of a character vector into a matrix
}
\note{
Updated 2023-09-21.
}
\examples{
x <- c(
    "TSPAN6_ENST00000373020",
    "TSPAN6_ENST00000494424",
    "TSPAN6_ENST00000496771",
    "TSPAN6_ENST00000612152",
    "TSPAN6_ENST00000614008"
)
x <- strSplit(x = x, split = "_")
print(x)
}
\seealso{
\itemize{
\item \code{\link[=strsplit]{strsplit()}}.
\item \code{\link[stringi:stri_split]{stringi::stri_split_fixed()}}.
\item \code{\link[stringr:str_split]{stringr::str_split_fixed()}}.
}
}
